+++ app/sql/data/album_count.sql
