+++ app/sql/data/artist_count.sql
