+++ app/sql/data/artist_total_time.sql
