+++ app/sql/data/get_artist_name.sql
