+++ app/sql/data/insert_cover_art.sql
@@ -1,0 +1,4 @@
+UPDATE albums
+SET cover_art_url = ?
+WHERE
+	spotify_id = ?;
