+++ app/sql/data/top_artist_albums.sql
