+++ app/sql/data/top_artists.sql
