+++ app/sql/data/top_played_artists.sql
