+++ app/static/main.js
