+++ app/static/style-redo.css
@@ -1,0 +1,185 @@
+@font-face {
+    font-family:cyberpunkwaifus;
+    src:url(../CyberpunkWaifus.woff) format('woff');
+    font-weight:400;
+    font-style:normal
+  }
+
+html {
+    height: 100%;
+    width: 100%;
+}
+
+body {
+    /*font-family: 'IBM Plex Mono', sans-serif;*/
+    /* font-family:cyberpunkwaifus; */
+    font-family: Arial, Helvetica, sans-serif;
+    color: white;
+    background-color: black;
+    opacity: 0.9;
+    width: 99%;
+    height: 100%;
+    font-size: 19px;
+}
+
+.TopTen {
+    border-style: solid;
+    border-radius: 20px;
+    border-width: 5pt;
+    border-color: white;
+    padding: 20px;
+}
+
+h1, h2, h3 {
+    font-family:cyberpunkwaifus;
+}
+
+.TitleBlock {
+    font-weight: bold;
+    font-size: 24px;
+    padding: 15px;
+    border-style: solid;
+    border-color: white;
+    margin: 5px;
+    display:inline-block;
+}
+
+.SettingBlock {
+    font-weight: bold;
+    font-size: 24px;
+    padding: 15px;
+    border-style: solid;
+    border-color: white;
+    margin: 5px auto;
+    /* text-align: center; */
+    /* display:inline-block; */
+    /* right: 50%; */
+    width: 25%;
+}
+
+.BigTitleBlock {
+    font-weight: bold;
+    font-size: 30px;
+    padding: 15px;
+    border-style: dashed;
+    border-color: white;
+    margin: 5px;
+    text-align: center;
+}
+
+.Graphs {
+    text-align: center;
+    padding: 5px;
+}
+
+.Graphs header {
+    text-align: center;
+}
+
+.Graph {
+    display: inline-block;
+    padding: 15px;
+}
+
+.Summary {
+    text-align: center;
+}
+
+.SummaryBlock {
+    /* border-style: solid;
+    border-radius: 40px; */
+    padding: 15px;
+    display: inline-block;
+    text-align: left;
+    /* margin: 15px; */
+}
+
+.SummaryBlock h2 {
+    font-size: 30px;
+    font-weight: 900;
+    padding-top: 10pt;
+}
+
+.SummaryBlock img {
+    height: 30px;
+    width: 30px;
+    padding: 0 0;
+    margin: 0 0;
+    margin-right: 10px;
+    border: 0 0;
+    line-height: 0;
+    font-size: 0;
+}
+
+.UserName {
+    float: right;
+    margin-top: 16px;
+    margin-right: 5px;
+}
+
+.UserPfp {
+    overflow: hidden;
+    float: right;
+    width: 48px;
+    height: 48px;
+    border-radius: 50%;
+}
+
+.UserPfp img {
+    width: auto;
+    height: 100%;
+}
+
+
+
+/* #Body div.Header2 {
+    padding-top: 10pt;
+    font-size: 18pt;
+}
+
+#Body div.Header2_u {
+    padding-top: 10pt;
+    font-size: 18pt;
+    text-decoration: underline;
+}
+
+#body div.TotalTime {
+    text-decoration: underline;
+} */
+
+a {
+    color: #ed145b;
+    text-align: center;
+    text-decoration: none;
+}
+
+a:hover {
+    color: #8b8fc6;
+}
+
+.selector {
+    color: #ed145b;
+    text-align: center;
+    text-decoration: none;
+}
+
+.selector:hover {
+    color: #8b8fc6;
+}
+
+
+.Artist {
+    text-align: center;
+}
+
+.Block {
+    width: 40%;
+    display: inline-block;
+    vertical-align: text-top;
+    text-align: left;
+    border-style: solid;
+    border-color: white;
+    padding: 10px;
+    margin: 20px;
+    margin-inline: 40px;
+}
