+++ app/templates/albums_month_overview.html
