+++ app/templates/artist.html
